requirements_dev.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cramjam==2.5.0
cycler==0.11.0
dateparser==1.1.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fastparquet==0.8.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
idna==3.3
imbalanced-learn==0.9.1
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pandas-ta==0.3.14b0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyarrow==8.0.0
pycparser==2.21
Pygments==2.12.0
pymongo==4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.0.0
regex==2022.3.2
requests==2.27.1
scikit-learn==1.1.1
scikit-optimize==0.9.0
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.2.2.post1
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.3
xgboost==1.6.1
yarl==1.7.2
zipp==3.8.0
